Add my local project
at this moment of the project I created the structure of the site with react and css ( I used react-router for the pages of the site ), with redux i create the store with all the pets inside and I show them on the different pages divided in categories. I created also a search bar for filter the pets by name.